requirements.txt
appdirs==1.4.4
asgiref==3.2.8
astroid==2.4.2
beautifulsoup4==4.8.2
black==20.8b1
branca==0.4.1
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
defusedxml==0.7.0rc1
Django==3.0.7
django-crontab==0.7.1
django-grappelli==2.14.2
djangorestframework==3.12.2
flake8==3.8.4
folium==0.11.0
idna==2.8
isort==5.6.4
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Lid==1.0.4
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.1
oauthlib==3.1.0
pathspec==0.8.1
Pillow==7.2.0
pyaml==20.4.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
pyTelegramBotAPI==3.6.6
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==1.9.5
sqlparse==0.3.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.7
wrapt==1.12.1
